+++ odoo/local-src/specific_shipping_costs/models/sale_order.py
@@ -2,14 +2,14 @@
 # Copyright 2018 Camptocamp SA
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html)
 
-from odoo import _, api, fields, models
+from odoo import api, fields, models
 
 
 class SaleOrder(models.Model):
     _inherit = 'sale.order'
 
     used_for_delivery_fee = fields.Boolean(
-        'Has been used for delivery fee calculation'
+        'Has been used for delivery fee calculation', copy=False
     )
 
     @api.model
@@ -43,20 +43,5 @@ class SaleOrder(models.Model):
         round_saleorders = round_saleorders.filtered(
             lambda r: r.carrier_id == carrier
         ).sorted(key=lambda r: r.id, reverse=True)[0]
-        round_saleorders.sudo().write(
-            {
-                'order_line': [
-                    (
-                        0,
-                        0,
-                        {
-                            'name': _('Shipping cost'),
-                            'product_id': carrier.product_id.id,
-                            'product_uom_qty': 1,
-                            'price_unit': carrier.fixed_price,
-                            'is_delivery': True,
-                        },
-                    )
-                ]
-            }
-        )
+        for order in round_saleorders.sudo():
+            order._create_delivery_line(carrier, carrier.fixed_price)